Refactor: unify SearchResult class, remove duplicate, resolve circular import, and update all usages
