+++ botnim/vector_store/__init__.py
@@ -1,5 +1,5 @@
 from .vector_store_openai import VectorStoreOpenAI
 from .vector_store_es import VectorStoreES
-from .search_config import SearchModeConfig, SearchResult
+from .search_config import SearchModeConfig
 
-__all__ = [VectorStoreOpenAI, VectorStoreES, SearchModeConfig, SearchResult]
+__all__ = [VectorStoreOpenAI, VectorStoreES, SearchModeConfig]
